feat: StudyMaking 페이지 관련 로직 useStudyMaking 훅 분리
